src/avx512/ed25519_avx512.rs
#[cfg(target_feature = "avx512f")]
use std::arch::x86_64::*;


// use curve25519_dalek::constants;
#[cfg(target_feature = "avx512f")]
use curve25519_dalek::scalar::Scalar;

/// the order of the Ed25519 subgroup (l), represented as a Scalar.
/// this constant is used in the Ed25519 signature scheme for use
/// particularly in modular arithmetic operations.
#[cfg(target_feature = "avx512f")]
const L: Scalar = Scalar::from_bits([
    0xed, 0xd3, 0xf5, 0x5c, 0x1a, 0x63, 0x12, 0x58, 0xd6, 0x9c, 0xf7, 0xa2, 0xde, 0xf9, 0xde, 0x14,
    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10,
]);

#[cfg(target_feature = "avx512f")]
pub unsafe fn ed25519_sign_avx512(
    private_key: &Vec<[u8; 32]>,
    message: &Vec<[[u8; 32]; 16]>,
) -> Vec<[[u8; 64]; 16]> {
    // step 1: hash the private key
    let mut h = Sha512::new();
    h.update(private_key);
    let hash = h.finalize();

    // step 2: compute scalar and prefix
    let scalar_bytes = &hash[0..32];
    let prefix = &hash[32..];

    // step 3: compute public key (we'll use curve25519_dalek for this part)
    let scalar = Scalar::from_bits(clamp_scalar(scalar_bytes));
    let public_key = (&constants::ED25519_BASEPOINT_TABLE * &scalar).compress();

    // step 4: compute r = H(prefix || message)
    let mut h = Sha512::new();
    h.update(prefix);
    h.update(message);
    let r = Scalar::from_hash(h);

    // step 5: compute R = rB
    let r_point = (&constants::ED25519_BASEPOINT_TABLE * &r).compress();

    // step 6: compute h = H(R || A || message)
    let mut h = Sha512::new();
    h.update(r_point.as_bytes());
    h.update(public_key.as_bytes());
    h.update(message);
    let h = Scalar::from_hash(h);

    // step 7: compute s = r + h * scalar (mod L)
    let s = r + h * scalar;

    // step 8: construct signature
    let mut signature = [0u8; 64];
    signature[..32].copy_from_slice(r_point.as_bytes());
    signature[32..].copy_from_slice(s.as_bytes());

    signature
}

#[cfg(target_feature = "avx512f")]
unsafe fn clamp_scalar(scalar: &[u8; 32]) -> [u8; 32] {
    let mut clamped = *scalar;
    clamped[0] &= 248;
    clamped[31] &= 63;
    clamped[31] |= 64;
    clamped
}

// avx-512 optimized field arithmetic operations
#[cfg(target_feature = "avx512f")]
mod field_ops {
    use super::*;

    pub unsafe fn field_mul_avx512(a: __m512i, b: __m512i) -> __m512i {
        // implement field multiplication using avx-512 instructions
        unimplemented!("Implement field multiplication using avx-512")
    }

    pub unsafe fn field_add_avx512(a: __m512i, b: __m512i) -> __m512i {
        _mm512_add_epi64(a, b)
    }

    pub unsafe fn scalar_mul_avx512(scalar: __m512i, point: __m512i) -> __m512i {
        // implement scalar multiplication using avx-512 instructions
        unimplemented!("Implement scalar multiplication using avx-512")
    }
}

// helper functions for avx-512 optimized ed25519 operations
#[cfg(target_feature = "avx512f")]
mod ed25519_ops {
    use super::*;
    use field_ops::*;

    // compute public key using avx-512
    pub unsafe fn compute_public_key_avx512(scalar: __m512i) -> __m512i {
        let base_point = load_ed25519_basepoint();
        scalar_mul_avx512(scalar, base_point)
    }

    // sign message using avx-512
    pub unsafe fn sign_message_avx512(
        scalar: __m512i,
        prefix: __m512i,
        message: &[u8],
    ) -> (__m512i, __m512i) {
        let r = compute_r_avx512(prefix, message);
        let r_point = scalar_mul_avx512(r, load_ed25519_basepoint());
        let public_key = compute_public_key_avx512(scalar);
        let h = compute_h_avx512(r_point, public_key, message);
        let s = compute_s_avx512(r, h, scalar);
        (r_point, s)
    }

    // helper function to load ed25519 base point
    unsafe fn load_ed25519_basepoint() -> __m512i {
        // steps:
        // 1. define the ed25519 base point coordinates
        // 2. load these coordinates into avx-512 registers
        // 3. return the loaded base point as a __m512i
    }

    // compute r = h(prefix || message)
    unsafe fn compute_r_avx512(prefix: __m512i, message: &[u8]) -> __m512i {
        // steps:
        // 1. concatenate prefix and message
        // 2. implement avx-512 optimized sha512 hash function
        // 3. hash the concatenated data
        // 4. reduce the 64-byte hash to a 32-byte scalar using reduce_scalar_avx512
        // 5. return the reduced scalar as __m512i
    }

    // compute h = h(r || a || m)
    unsafe fn compute_h_avx512(r_point: __m512i, public_key: __m512i, message: &[u8]) -> __m512i {
        // steps:
        // 1. concatenate r_point, public_key, and message
        // 2. use the same avx-512 optimized sha512 hash function as in compute_r_avx512
        // 3. hash the concatenated data
        // 4. reduce the 64-byte hash to a 32-byte scalar using reduce_scalar_avx512
        // 5. return the reduced scalar as __m512i
    }

    // compute s = r + h * scalar
    unsafe fn compute_s_avx512(r: __m512i, h: __m512i, scalar: __m512i) -> __m512i {
        let h_scalar = scalar_mul_avx512(h, scalar);
        field_add_avx512(r, h_scalar)
    }

    // reduce 64-byte hash to 32-byte scalar
    unsafe fn reduce_scalar_avx512(hash: __m512i) -> __m512i {
        // steps:
        // 1. implement the scalar reduction algorithm for ed25519
        // 2. use avx-512 instructions to perform the reduction efficiently
        // 3. ensure the result is a valid ed25519 scalar (less than the group order)
        // 4. return the reduced scalar as __m512i
    }
}
